# backend/docs/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.1.0
  title: HomiesHomes Real Estate API
  description: A RESTful API for HomiesHomes real estate listings platform. This API provides endpoints for managing users, properties, bookings, favorites, and agent requests.
  contact:
    name: Mandeep Pandya
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://gammacairo-deltareward-9000.codio-box.uk"
    description: Codio API server
  - url: "https://gammacairo-deltareward-3000.codio-box.uk"
    description: Codio frontend server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Authentication. A valid JWT token must be passed in the Authorization header.
        Example: `Authorization: Bearer {token}`

        You can obtain a token by using the /login endpoint with valid credentials.
        Tokens are valid for 1 hour after issuance.
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed"
paths:
...
  /favourites:
    get:
      tags:
        - Favourites
      summary: Get user favourites
      description: Returns a list of all favourited properties for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of favourited properties
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    property_id:
                      type: integer
                      example: 1
                    user_id:
                      type: integer
                      example: 3
                    favourite_date:
                      type: string
                      format: date-time
                      example: "2023-04-01T10:30:00Z"
                    title:
                      type: string
                      example: "Modern Apartment Downtown"
                    price:
                      type: number
                      example: 250000
                    location:
                      type: string
                      example: "Birmingham, UK"
                    description:
                      type: string
                      example: "Beautiful modern apartment in the heart of the city."
                    property_type:
                      type: string
                      example: "Apartment"
                    bedrooms:
                      type: integer
                      example: 2
                    bathrooms:
                      type: integer
                      example: 1
                    image_url:
                      type: string
                      example: "prop1.jpg"
                    agent_id:
                      type: integer
                      example: 2
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Favourites
      summary: Add favourite
      description: Adds a property to the user's favourites.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - property_id
              properties:
                property_id:
                  type: integer
                  example: 1
      responses:
        "201":
          description: Property successfully added to favourites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Property added to favourites"
                  favorite:
                    type: object
                    properties:
                      user_id:
                        type: integer
                        example: 3
                      property_id:
                        type: integer
                        example: 1
        "400":
          description: Bad request - property already in favourites
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Property already in favourites"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found - property does not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Property not found"
  /favourites/{property_id}:
    parameters:
      - in: path
        name: property_id
        schema:
          type: integer
        required: true
        description: Property ID to remove from favourites

    delete:
      tags:
        - Favourites
      summary: Remove favourite
      description: Removes a property from the user's favourites.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Property successfully removed from favourites
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Property removed from favourites"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Favourite not found"
...